identifiers: swap to 64bit identifier
Swap to a 64bit identifier to eliminate any chance that we will run out
of addresses and give us more freedom in assigning sequential IDs.

+++ lib/identifiers.c
@@ -8,7 +8,7 @@
 
 #include <infuse/identifiers.h>
 
-uint32_t infuse_device_id(void)
+uint64_t infuse_device_id(void)
 {
 	return 1234;
 }
